backend/services/requirements.txt
uvicorn>=0.22.0,<1.0.0
pydantic>=1.10.13,<2.0.0
sqlalchemy>=1.4.52,<2.0.0
databases>=0.8.0,<1.0.0
aiofiles>=0.8.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
bcrypt>=4.0.1,<5.0.0
pytest>=7.2.0,<8.0.0
httpx>=0.23.0,<1.0.0
pymongo>=4.2.0,<5.0.0
python-dotenv>=0.21.0,<1.0.0
requests>=2.31.0,<3.0.0
biopython>=1.83,<2.0.0
numpy>=1.26.4,<2.0.0
pandas>=2.2.0,<3.0.0
pyvcf3>=1.0.3,<2.0.0
# pysam>=0.22.0,<1.0.0 # pysam is not available on Windows (might move to WSL or install via conda)